Refactor resources subpackage a bit
