+++ pygfx/resources/__init__.py
@@ -17,7 +17,8 @@ In pygfx, data is stored in buffers and textures. We collectively call these res
 
 # flake8: noqa
 
-from ._buffer import Resource, Buffer
+from ._base import Resource
+from ._buffer import Buffer
 from ._texture import Texture
 
 __all__ = ["Resource", "Buffer", "Texture"]
